Top half of the map
The bags of flour/sugar, bottle of vanilla, and stick of butter are now
in the house. You cannot pick them up unless you find out what they
really are!

ex: You won't pick up the bar of soap, but once you look and see it's a
stick of butter you can definitely take it.